+++ ls4/ReuestHandler.js
@@ -2,7 +2,7 @@ const { createWriteStream, createReadStream } = require("fs");
 const url = require("url");
 const path = require("path");
 const { allowedContentTypes } = require("./settings");
-const { parseMessage } = require("./utils");
+const { parseMessage, transformData } = require("./utils");
 
 
 
@@ -43,46 +43,21 @@ exports.Request = {
                 let result = JSON.parse(storedData);
 
                 Object.keys(parseRequestParams).map(handler => {
-                    if (par.query[handler]) {
-                        result = parseRequestParams[handler](result, par.query[handler]);
+                    const params = par.query[handler];
+                    if (params) {
+                        result = parseRequestParams[handler](result, params);
                     }
                 })
-
-                // sortMessages(result)
-                 console.log(result);   
                 res.end(JSON.stringify(result));
-
             }
             res.end(messages.toString())
         })
     }
 }
 
-const sortMessages = (messages) => {
-    console.log('sort', messages);
-
-    const result = messages.sort((a, b) => {
-        return (a.text < b.text) ? -1 : 1  
-    })
-    return result;
-}
-
-const limitMessage = (messages, limitValue) => {
-    console.log('limit', messages);
-    const result = messages.slice(0, limitValue);
-    return result;
-}
-
-const skipMessage = (messages, skipValue) => {
-    debugger
-    console.log('skip', messages);
-    const result = messages.slice(Number(skipValue) + 1);
-    return result;
-}
-
 const parseRequestParams = {
-    sort: (data) => sortMessages(data),
-    skip: (data, param) => skipMessage(data, param),
-    limit: (data, param) => limitMessage(data, param)
+    sort: (data) => transformData.sortMessages(data),
+    skip: (data, param) => transformData.skipMessage(data, param),
+    limit: (data, param) => transformData.limitMessage(data, param)
 }
